044-update-Change screenSizeInch to double
make sure that our Databases using double if not using
--ALTER TABLE sale_item MODIFY screenSizeInch DOUBLE;

+++ BE/itb-mshop-be/src/main/java/org/example/itbmshopbe/dtos/SaleItemDetailDto.java
@@ -17,7 +17,7 @@ public class SaleItemDetailDto {
     private String description;
     private Integer price;
     private Integer ramGb;
-    private BigDecimal screenSizeInch;
+    private Double screenSizeInch;
     private Integer quantity;
     private Integer storageGb;
     private String color;
